Do some refactoring
The files to import have been moved to the resources folder outside the java dir, to match with Maven requirements.
Also the methods to import them has been changed slightly because it didn't work if the file was moved.
Little issues with the document verification and with the input of the user have been fixed.